src/requirements.txt
sdbmigrate[postgres]==1.1.0
aiohttp
aiohttp_swagger3
pyyaml==5.4.1
asyncpg==0.28.0
prometheus-client==0.17.1
